# loser-tree: add sequence abstraction