--- a/loser/loser.go
+++ b/loser/loser.go
@@ -4,111 +4,157 @@ package loser
 
 import "golang.org/x/exp/constraints"
 
-func New[E constraints.Ordered](lists [][]E, maxVal E) *Tree[E] {
-	nLists := len(lists)
-	t := Tree[E]{
+type Value constraints.Ordered
+
+type Sequence[E Value] interface {
+	At() E      // Returns the current value.
+	Next() bool // Advances and returns true if there is a value at this new position.
+}
+
+func New[E Value, S Sequence[E]](sequences []S, maxVal E) *Tree[E, S] {
+	nSequences := len(sequences)
+	t := Tree[E, S]{
 		maxVal: maxVal,
-		nodes:  make([]node[E], nLists*2),
+		nodes:  make([]node[E, S], nSequences*2),
 	}
-	for i, s := range lists {
-		t.nodes[i+nLists].items = s
-		t.moveNext(i + nLists) // Must call Next on each item so that At() has a value.
+	for i, s := range sequences {
+		t.nodes[i+nSequences].items = s
+		t.moveNext(i + nSequences) // Must call Next on each item so that At() has a value.
 	}
-	if nLists > 0 {
+	if nSequences > 0 {
 		t.nodes[0].index = -1 // flag to be initialized on first call to Next().
 	}
 	return &t
 }
 
+// Call the close function on all sequences that are still open.
+func (t *Tree[E, S]) Close() {
+	for _, e := range t.nodes[len(t.nodes)/2 : len(t.nodes)] {
+		if e.index == -1 {
+			continue
+		}
+	}
+}
+
 // A loser tree is a binary tree laid out such that nodes N and N+1 have parent N/2.
 // We store M leaf nodes in positions M...2M-1, and M-1 internal nodes in positions 1..M-1.
 // Node 0 is a special node, containing the winner of the contest.
-type Tree[E constraints.Ordered] struct {
+type Tree[E Value, S Sequence[E]] struct {
 	maxVal E
-	nodes  []node[E]
+	nodes  []node[E, S]
 }
 
-type node[E constraints.Ordered] struct {
+type node[E Value, S Sequence[E]] struct {
 	index int // This is the loser for all nodes except the 0th, where it is the winner.
 	value E   // Value copied from the loser node, or winner for node 0.
-	items []E // Only populated for leaf nodes.
+	items S   // Only populated for leaf nodes.
 }
 
-func (t *Tree[E]) moveNext(index int) bool {
+func (t *Tree[E, S]) moveNext(index int) bool {
 	n := &t.nodes[index]
-	if len(n.items) > 0 {
-		n.value = n.items[0]
-		n.items = n.items[1:]
-		return true
-	}
-	n.value = t.maxVal
-	n.index = -1
-	return false
-}
-
-func (t *Tree[E]) Winner() E {
-	return t.nodes[t.nodes[0].index].value
-}
-
-func (t *Tree[E]) Next() bool {
-	if len(t.nodes) == 0 {
-		return false
-	}
-	if t.nodes[0].index == -1 { // If tree has not been initialized yet, do that.
-		t.initialize()
-		return t.nodes[t.nodes[0].index].index != -1
-	}
-	if t.nodes[t.nodes[0].index].index == -1 { // already exhausted
-		return false
-	}
-	if t.moveNext(t.nodes[0].index) {
-		t.replayGames(t.nodes[0].index)
+	ret := n.items.Next()
+	if ret {
+		n.value = n.items.At()
 	} else {
-		t.sequenceEnded(t.nodes[0].index)
+		n.value = t.maxVal
+		n.index = -1
 	}
-	return t.nodes[t.nodes[0].index].index != -1
+	return ret
 }
 
-func (t *Tree[E]) initialize() {
-	winners := make([]int, len(t.nodes))
-	// Initialize leaf nodes as winners to start.
-	for i := len(t.nodes) / 2; i < len(t.nodes); i++ {
-		winners[i] = i
-	}
-	for i := len(t.nodes) - 2; i > 0; i -= 2 {
-		// At each stage the winners play each other, and we record the loser in the node.
-		loser, winner := t.playGame(winners[i], winners[i+1])
-		p := parent(i)
-		t.nodes[p].index = loser
-		t.nodes[p].value = t.nodes[loser].value
-		winners[p] = winner
-	}
-	t.nodes[0].index = winners[1]
-	t.nodes[0].value = t.nodes[winners[1]].value
+func (t *Tree[E, S]) Winner() S {
+	return t.nodes[t.nodes[0].index].items
 }
 
-// Starting at pos, which is a winner, re-consider all values up to the root.
-func (t *Tree[E]) replayGames(pos int) {
+func (t *Tree[E, S]) At() E {
+	return t.nodes[0].value
+}
+
+func (t *Tree[E, S]) Next() bool {
+	nodes := t.nodes
+	if len(nodes) == 0 {
+		return false
+	}
+	if nodes[0].index == -1 { // If tree has not been initialized yet, do that.
+		t.initialize()
+		return nodes[nodes[0].index].index != -1
+	}
+	if t.moveNext(nodes[0].index) {
+		t.replayGames(nodes[0].index)
+	} else {
+		t.sequenceEnded(nodes[0].index)
+	}
+	return nodes[nodes[0].index].index != -1
+}
+
+// Current winner has been advanced independently; fix up the loser tree.
+func (t *Tree[E, S]) Fix(closed bool) {
+	nodes := t.nodes
+	cur := &nodes[nodes[0].index]
+	if closed {
+		cur.value = t.maxVal
+		cur.index = -1
+	} else {
+		cur.value = cur.items.At()
+	}
+	t.replayGames(nodes[0].index)
+}
+
+func (t *Tree[E, S]) IsEmpty() bool {
+	nodes := t.nodes
+	if nodes[0].index == -1 { // If tree has not been initialized yet, do that.
+		t.initialize()
+	}
+	return nodes[nodes[0].index].index == -1
+}
+
+func (t *Tree[E, S]) initialize() {
+	winner := t.playGame(1)
+	t.nodes[0].index = winner
+	t.nodes[0].value = t.nodes[winner].value
+}
+
+// Find the winner at position pos; if it is a non-leaf node, store the loser.
+// pos must be >= 1 and < len(t.nodes)
+func (t *Tree[E, S]) playGame(pos int) int {
+	nodes := t.nodes
+	if pos >= len(nodes)/2 {
+		return pos
+	}
+	left := t.playGame(pos * 2)
+	right := t.playGame(pos*2 + 1)
+	var loser, winner int
+	if nodes[left].value < nodes[right].value {
+		loser, winner = right, left
+	} else {
+		loser, winner = left, right
+	}
+	nodes[pos].index = loser
+	nodes[pos].value = nodes[loser].value
+	return winner
+}
+
+// Starting at pos, re-consider all values up to the root.
+func (t *Tree[E, S]) replayGames(pos int) {
+	nodes := t.nodes
 	// At the start, pos is a leaf node, and is the winner at that level.
-	n := parent(pos)
-	for n != 0 {
-		// If n.value < pos.value then pos loses.
-		// If they are equal, pos wins because n could be a sequence that ended, with value maxval.
-		if t.nodes[n].value < t.nodes[pos].value {
-			loser := pos
+	winningValue := nodes[pos].value
+	for n := parent(pos); n != 0; n = parent(n) {
+		node := &nodes[n]
+		if node.value < winningValue {
 			// Record pos as the loser here, and the old loser is the new winner.
-			pos = t.nodes[n].index
-			t.nodes[n].index = loser
-			t.nodes[n].value = t.nodes[loser].value
+			node.index, pos = pos, node.index
+			node.value, winningValue = winningValue, node.value
 		}
-		n = parent(n)
 	}
 	// pos is now the winner; store it in node 0.
-	t.nodes[0].index = pos
-	t.nodes[0].value = t.nodes[pos].value
+	nodes[0].index = pos
+	nodes[0].value = winningValue
 }
 
-func (t *Tree[E]) sequenceEnded(pos int) {
+func parent(i int) int { return i >> 1 }
+
+func (t *Tree[E, S]) sequenceEnded(pos int) {
 	// Find the first active sequence which used to lose to it.
 	n := parent(pos)
 	for n != 0 && t.nodes[t.nodes[n].index].index == -1 {
@@ -129,17 +175,8 @@ func (t *Tree[E]) sequenceEnded(pos int) {
 	t.replayGames(winner)
 }
 
-func (t *Tree[E]) playGame(a, b int) (loser, winner int) {
-	if t.nodes[a].value < t.nodes[b].value {
-		return b, a
-	}
-	return a, b
-}
-
-func parent(i int) int { return i / 2 }
-
 // Add a new list to the merge set
-func (t *Tree[E]) Push(list []E) {
+func (t *Tree[E, S]) Push(list S) {
 	// First, see if we can replace one that was previously finished.
 	for newPos := len(t.nodes) / 2; newPos < len(t.nodes); newPos++ {
 		if t.nodes[newPos].index == -1 {
@@ -156,7 +193,7 @@ func (t *Tree[E]) Push(list []E) {
 		size *= 2
 	}
 	newPos := size + len(t.nodes)/2
-	newNodes := make([]node[E], size*2)
+	newNodes := make([]node[E, S], size*2)
 	// Copy data over and fix up the indexes.
 	for i, n := range t.nodes[len(t.nodes)/2:] {
 		newNodes[i+size] = n